Add timeout logic
There seems to be a bug in the tiered list implementation
however, it says that the list is empty after
a single try, even though there are multiple tiers.

The way the logic works now, it will always wait 1s, even
if the announce has arrived earlier. We should probably race.