Verify hashes in piece-writer
This makes it so that the piece-writer will verify the
sha1 hashes of the pieces, and then update our current
piece sets, as well as inform the peers of this piece
arrival, and then finally write the pieces to disk.
This improves performance a bit, since peers no longer
wait for slow disk writing in order to continue downloading
from their end.

Note that this does not inform peers of bad pieces, which should be
done.